Renaming renderer -> browser in jekyll-bookshop
