+++ generator-plugins/jekyll/jekyll-bookshop/lib/jekyll-bookshop.rb
@@ -5,14 +5,14 @@ require "pathname"
 require "dry/inflector"
 
 require_relative "jekyll-bookshop/tags/bookshop-tag"
-require_relative "jekyll-bookshop/tags/renderer-tag"
+require_relative "jekyll-bookshop/tags/browser-tag"
 require_relative "jekyll-bookshop/tags/style-tag"
 require_relative "jekyll-bookshop/init-styles"
 require_relative "jekyll-bookshop/site-data"
 
 Liquid::Template.register_tag("bookshop", JekyllBookshop::Tag)
 Liquid::Template.register_tag("bookshop_scss", JekyllBookshop::StyleTag)
-Liquid::Template.register_tag("bookshop_renderer", JekyllBookshop::RendererTag)
+Liquid::Template.register_tag("bookshop_browser", JekyllBookshop::BrowserTag)
 
 Jekyll::Hooks.register :site, :after_init do |site|
   JekyllBookshop::Styles.open_bookshop(site)
