+++ generator-plugins/jekyll/jekyll-bookshop/lib/jekyll-bookshop/site-data.rb
@@ -21,7 +21,7 @@ module JekyllBookshop
 
       @site.data["_bookshop_site_data"] = { "site" => payload_collections }
 
-      update_renderer_pages
+      update_browser_pages
 
       Jekyll.logger.info "Bookshop:",
                          "Bookshop site data generated"
@@ -40,13 +40,13 @@ module JekyllBookshop
       })
     end
 
-    def self.update_renderer_pages
+    def self.update_browser_pages
       @site.pages.each do |page|
         next unless @site.data["_bookshop_data_pages"]&.include?(page.url)
 
         page.output = page.output.gsub(
-          %r!bookshop_renderer_site_data = null!,
-          "bookshop_renderer_site_data = #{@site.data["_bookshop_site_data"].to_json}"
+          %r!bookshop_browser_site_data = null!,
+          "bookshop_browser_site_data = #{@site.data["_bookshop_site_data"].to_json}"
         )
       end
       @site.data["_bookshop_data_pages"] = nil
