+++ generator-plugins/jekyll/jekyll-bookshop/spec/fixture_site/index.html
@@ -8,4 +8,4 @@ blocks:
 
 {% bookshop item title="Item Test" %}
 
-{% bookshop_renderer :1234 %}
+{% bookshop_browser localhost:1234/bookshop.js %}
